Added PublicationsJob and refactored services
