+++ src/Core/Publications/Services/Contracts/IPublicationService.cs
@@ -1,4 +1,5 @@
-﻿using Biobanks.Publications.Core.Services.Dto;
+﻿using Biobanks.Entities.Data;
+using Biobanks.Publications.Core.Services.Dto;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -6,6 +7,8 @@ namespace Biobanks.Publications.Core.Services.Contracts
 {
     public interface IPublicationService
     {
+        Task<IEnumerable<Publication>> ListOrganisationPublications(int organisationId);
+
         Task AddOrganisationPublications(int organisationId, IEnumerable<PublicationDto> publications);
     }
 }
